+++ tests/test_logging.py
@@ -1,6 +1,6 @@
 import numpy as np
 
-from eulerpi.core.data_transformation import DataIdentity
+from eulerpi.core.data_transformations import DataIdentity
 from eulerpi.core.kde import calc_kernel_width
 from eulerpi.core.model import Model
 from eulerpi.core.transformations import evaluate_density
