+++ tests/test_transformations.py
@@ -2,7 +2,7 @@ import jax.numpy as jnp
 import numpy as np
 import pytest
 
-from eulerpi.core.data_transformation import DataIdentity
+from eulerpi.core.data_transformations import DataIdentity
 from eulerpi.core.kde import calc_kernel_width, eval_kde_gauss
 from eulerpi.core.model import ArtificialModelInterface, JaxModel
 from eulerpi.core.transformations import calc_gram_determinant
